// website-next/src/lib/client/types.gen.ts
export type ServerResponse = {
    id?: string;
    sv_hostname?: string;
    mapname?: string;
    g_gametype?: number;
    fraglimit?: number;
    timelimit?: number;
    sv_maxclients?: number;
    g_needpass?: number;
    capturelimit?: number;
    version?: string;
    location?: string;
    players?: number;
    ping?: number;
    host?: string;
    port?: number;
    challenge?: string;
    sv_maxPing?: number;
    sv_minPing?: number;
    com_gamename?: string;
    com_protocol?: number;
    dmflags?: number;
    sv_privateClients?: number;
    sv_minRate?: number;
    sv_maxRate?: number;
    sv_dlRate?: number;
    sv_floodProtect?: number;
    sv_allowDownload?: number;
    bot_minplayers?: number;
    gamename?: string;
    g_maxGameClients?: number;
    users?: Array<ServerUserResponse>;
    proxyPort?: number;
    targetPort?: number;
};

export type ServerUserResponse = {
    score?: number;
    ping?: number;
    name?: string;
};